refactor: remove unnecessary ref for pendingRemovals
